// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type Barter = {
  __typename?: 'Barter';
  id: Scalars['ID'];
  level: Scalars['Int'];
  requiredItems: Array<Maybe<ContainedItem>>;
  /** @deprecated Use level instead. */
  requirements: Array<Maybe<PriceRequirement>>;
  rewardItems: Array<Maybe<ContainedItem>>;
  /** @deprecated Use trader and level instead. */
  source: Scalars['String'];
  /** @deprecated Use trader instead. */
  sourceName: ItemSourceName;
  taskUnlock?: Maybe<Task>;
  trader: Trader;
};
// ...
export type ContainedItem = {
  __typename?: 'ContainedItem';
  attributes?: Maybe<Array<Maybe<ItemAttribute>>>;
  count: Scalars['Float'];
  item: Item;
  quantity: Scalars['Float'];
};

export type Craft = {
  __typename?: 'Craft';
  duration: Scalars['Int'];
  id: Scalars['ID'];
  level: Scalars['Int'];
  requiredItems: Array<Maybe<ContainedItem>>;
  requiredQuestItems: Array<Maybe<QuestItem>>;
  /** @deprecated Use stationLevel instead. */
  requirements: Array<Maybe<PriceRequirement>>;
  rewardItems: Array<Maybe<ContainedItem>>;
  /** @deprecated Use stationLevel instead. */
  source: Scalars['String'];
  /** @deprecated Use stationLevel instead. */
  sourceName: Scalars['String'];
  station: HideoutStation;
  taskUnlock?: Maybe<Task>;
};
// ...
export type HideoutStation = {
  __typename?: 'HideoutStation';
  /** crafts is only available via the hideoutStations query. */
  crafts: Array<Maybe<Craft>>;
  id: Scalars['ID'];
  levels: Array<Maybe<HideoutStationLevel>>;
  name: Scalars['String'];
  normalizedName: Scalars['String'];
  tarkovDataId?: Maybe<Scalars['Int']>;
};
// ...
export type Item = {
  __typename?: 'Item';
  accuracyModifier?: Maybe<Scalars['Float']>;
  avg24hPrice?: Maybe<Scalars['Int']>;
  backgroundColor: Scalars['String'];
  bartersFor: Array<Maybe<Barter>>;
  bartersUsing: Array<Maybe<Barter>>;
  baseImageLink?: Maybe<Scalars['String']>;
  basePrice: Scalars['Int'];
  blocksHeadphones?: Maybe<Scalars['Boolean']>;
  /** @deprecated Use category instead. */
  bsgCategory?: Maybe<ItemCategory>;
  bsgCategoryId?: Maybe<Scalars['String']>;
  buyFor?: Maybe<Array<ItemPrice>>;
  categories: Array<Maybe<ItemCategory>>;
  category?: Maybe<ItemCategory>;
  /** @deprecated No longer meaningful with inclusion of Item category. */
  categoryTop?: Maybe<ItemCategory>;
  changeLast48h?: Maybe<Scalars['Float']>;
  changeLast48hPercent?: Maybe<Scalars['Float']>;
  conflictingItems?: Maybe<Array<Maybe<Item>>>;
  conflictingSlotIds?: Maybe<Array<Maybe<Scalars['String']>>>;
  containsItems?: Maybe<Array<Maybe<ContainedItem>>>;
  craftsFor: Array<Maybe<Craft>>;
  craftsUsing: Array<Maybe<Craft>>;
  description?: Maybe<Scalars['String']>;
  ergonomicsModifier?: Maybe<Scalars['Float']>;
  fleaMarketFee?: Maybe<Scalars['Int']>;
  gridImageLink?: Maybe<Scalars['String']>;
  /** @deprecated Fallback handled automatically by gridImageLink. */
  gridImageLinkFallback: Scalars['String'];
  handbookCategories: Array<Maybe<ItemCategory>>;
  hasGrid?: Maybe<Scalars['Boolean']>;
  height: Scalars['Int'];
  high24hPrice?: Maybe<Scalars['Int']>;
  /** historicalPrices is only available via the item and items queries. */
  historicalPrices?: Maybe<Array<Maybe<HistoricalPricePoint>>>;
  iconLink?: Maybe<Scalars['String']>;
  /** @deprecated Fallback handled automatically by iconLink. */
  iconLinkFallback: Scalars['String'];
  id: Scalars['ID'];
  image8xLink?: Maybe<Scalars['String']>;
  image512pxLink?: Maybe<Scalars['String']>;
  /** @deprecated Use inspectImageLink instead. */
  imageLink?: Maybe<Scalars['String']>;
  /** @deprecated Fallback handled automatically by inspectImageLink. */
  imageLinkFallback: Scalars['String'];
  inspectImageLink?: Maybe<Scalars['String']>;
  lastLowPrice?: Maybe<Scalars['Int']>;
  lastOfferCount?: Maybe<Scalars['Int']>;
  link?: Maybe<Scalars['String']>;
  loudness?: Maybe<Scalars['Int']>;
  low24hPrice?: Maybe<Scalars['Int']>;
  name?: Maybe<Scalars['String']>;
  normalizedName?: Maybe<Scalars['String']>;
  properties?: Maybe<ItemProperties>;
  receivedFromTasks: Array<Maybe<Task>>;
  recoilModifier?: Maybe<Scalars['Float']>;
  sellFor?: Maybe<Array<ItemPrice>>;
  shortName?: Maybe<Scalars['String']>;
  /** @deprecated Use sellFor instead. */
  traderPrices: Array<Maybe<TraderPrice>>;
  /** @deprecated Use the lang argument on queries instead. */
  translation?: Maybe<ItemTranslation>;
  types: Array<Maybe<ItemType>>;
  updated?: Maybe<Scalars['String']>;
  usedInTasks: Array<Maybe<Task>>;
  velocity?: Maybe<Scalars['Float']>;
  weight?: Maybe<Scalars['Float']>;
  width: Scalars['Int'];
  wikiLink?: Maybe<Scalars['String']>;
};


export type ItemFleaMarketFeeArgs = {
  count?: InputMaybe<Scalars['Int']>;
  hideoutManagementLevel?: InputMaybe<Scalars['Int']>;
  intelCenterLevel?: InputMaybe<Scalars['Int']>;
  price?: InputMaybe<Scalars['Int']>;
  requireAll?: InputMaybe<Scalars['Boolean']>;
};


export type ItemTranslationArgs = {
  languageCode?: InputMaybe<LanguageCode>;
};

export type ItemAttribute = {
  __typename?: 'ItemAttribute';
  name: Scalars['String'];
  type: Scalars['String'];
  value?: Maybe<Scalars['String']>;
};

export type ItemCategory = {
  __typename?: 'ItemCategory';
  children?: Maybe<Array<Maybe<ItemCategory>>>;
  id: Scalars['ID'];
  name: Scalars['String'];
  normalizedName: Scalars['String'];
  parent?: Maybe<ItemCategory>;
};
// ...
export type ItemPrice = {
  __typename?: 'ItemPrice';
  currency?: Maybe<Scalars['String']>;
  currencyItem?: Maybe<Item>;
  price?: Maybe<Scalars['Int']>;
  priceRUB?: Maybe<Scalars['Int']>;
  /** @deprecated Use vendor instead. */
  requirements: Array<Maybe<PriceRequirement>>;
  /** @deprecated Use vendor instead. */
  source?: Maybe<ItemSourceName>;
  vendor: Vendor;
};

export type ItemProperties = ItemPropertiesAmmo | ItemPropertiesArmor | ItemPropertiesArmorAttachment | ItemPropertiesBackpack | ItemPropertiesBarrel | ItemPropertiesChestRig | ItemPropertiesContainer | ItemPropertiesFoodDrink | ItemPropertiesGlasses | ItemPropertiesGrenade | ItemPropertiesHelmet | ItemPropertiesKey | ItemPropertiesMagazine | ItemPropertiesMedKit | ItemPropertiesMedicalItem | ItemPropertiesMelee | ItemPropertiesNightVision | ItemPropertiesPainkiller | ItemPropertiesPreset | ItemPropertiesScope | ItemPropertiesStim | ItemPropertiesSurgicalKit | ItemPropertiesWeapon | ItemPropertiesWeaponMod;

export type ItemPropertiesAmmo = {
  __typename?: 'ItemPropertiesAmmo';
  /** @deprecated Use accuracyModifier instead. */
  accuracy?: Maybe<Scalars['Int']>;
  accuracyModifier?: Maybe<Scalars['Float']>;
  ammoType?: Maybe<Scalars['String']>;
  armorDamage?: Maybe<Scalars['Int']>;
  caliber?: Maybe<Scalars['String']>;
  damage?: Maybe<Scalars['Int']>;
  durabilityBurnFactor?: Maybe<Scalars['Float']>;
  fragmentationChance?: Maybe<Scalars['Float']>;
  heatFactor?: Maybe<Scalars['Float']>;
  heavyBleedModifier?: Maybe<Scalars['Float']>;
  initialSpeed?: Maybe<Scalars['Float']>;
  lightBleedModifier?: Maybe<Scalars['Float']>;
  penetrationChance?: Maybe<Scalars['Float']>;
  penetrationPower?: Maybe<Scalars['Int']>;
  projectileCount?: Maybe<Scalars['Int']>;
  /** @deprecated Use recoilModifier instead. */
  recoil?: Maybe<Scalars['Float']>;
  recoilModifier?: Maybe<Scalars['Float']>;
  ricochetChance?: Maybe<Scalars['Float']>;
  stackMaxSize?: Maybe<Scalars['Int']>;
  tracer?: Maybe<Scalars['Boolean']>;
  tracerColor?: Maybe<Scalars['String']>;
};

export type ItemPropertiesArmor = {
  __typename?: 'ItemPropertiesArmor';
  class?: Maybe<Scalars['Int']>;
  durability?: Maybe<Scalars['Int']>;
  ergoPenalty?: Maybe<Scalars['Int']>;
  material?: Maybe<ArmorMaterial>;
  repairCost?: Maybe<Scalars['Int']>;
  speedPenalty?: Maybe<Scalars['Float']>;
  turnPenalty?: Maybe<Scalars['Float']>;
  zones?: Maybe<Array<Maybe<Scalars['String']>>>;
};

export type ItemPropertiesArmorAttachment = {
  __typename?: 'ItemPropertiesArmorAttachment';
  blindnessProtection?: Maybe<Scalars['Float']>;
  class?: Maybe<Scalars['Int']>;
  durability?: Maybe<Scalars['Int']>;
  ergoPenalty?: Maybe<Scalars['Int']>;
  headZones?: Maybe<Array<Maybe<Scalars['String']>>>;
  material?: Maybe<ArmorMaterial>;
  repairCost?: Maybe<Scalars['Int']>;
  speedPenalty?: Maybe<Scalars['Float']>;
  turnPenalty?: Maybe<Scalars['Float']>;
};
// ...
export type ItemPropertiesKey = {
  __typename?: 'ItemPropertiesKey';
  uses?: Maybe<Scalars['Int']>;
};
// ...
export enum ItemSourceName {
  Fence = 'fence',
  FleaMarket = 'fleaMarket',
  Jaeger = 'jaeger',
  Mechanic = 'mechanic',
  Peacekeeper = 'peacekeeper',
  Prapor = 'prapor',
  Ragman = 'ragman',
  Skier = 'skier',
  Therapist = 'therapist'
}

export type ItemStorageGrid = {
  __typename?: 'ItemStorageGrid';
  filters: ItemFilters;
  height: Scalars['Int'];
  width: Scalars['Int'];
};

/**
 * The below types are all deprecated and may not return current data.
 * ItemTranslation has been replaced with the lang argument on all queries
 */
export type ItemTranslation = {
  __typename?: 'ItemTranslation';
  /** @deprecated Use the lang argument on queries instead. */
  description?: Maybe<Scalars['String']>;
  /** @deprecated Use the lang argument on queries instead. */
  name?: Maybe<Scalars['String']>;
  /** @deprecated Use the lang argument on queries instead. */
  shortName?: Maybe<Scalars['String']>;
};

export enum ItemType {
  Ammo = 'ammo',
  AmmoBox = 'ammoBox',
  Any = 'any',
  Armor = 'armor',
  Backpack = 'backpack',
  Barter = 'barter',
  Container = 'container',
  Glasses = 'glasses',
  Grenade = 'grenade',
  Gun = 'gun',
  Headphones = 'headphones',
  Helmet = 'helmet',
  Injectors = 'injectors',
  Keys = 'keys',
  MarkedOnly = 'markedOnly',
  Meds = 'meds',
  Mods = 'mods',
  NoFlea = 'noFlea',
  PistolGrip = 'pistolGrip',
  Preset = 'preset',
  Provisions = 'provisions',
  Rig = 'rig',
  Suppressor = 'suppressor',
  Wearable = 'wearable'
}
// ...
export type PriceRequirement = {
  __typename?: 'PriceRequirement';
  stringValue?: Maybe<Scalars['String']>;
  type: RequirementType;
  value?: Maybe<Scalars['Int']>;
};
// ...
export type QuestItem = {
  __typename?: 'QuestItem';
  baseImageLink?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  gridImageLink?: Maybe<Scalars['String']>;
  height?: Maybe<Scalars['Int']>;
  iconLink?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  image8xLink?: Maybe<Scalars['String']>;
  image512pxLink?: Maybe<Scalars['String']>;
  inspectImageLink?: Maybe<Scalars['String']>;
  name: Scalars['String'];
  normalizedName?: Maybe<Scalars['String']>;
  shortName?: Maybe<Scalars['String']>;
  width?: Maybe<Scalars['Int']>;
};
// ...
export type Task = {
  __typename?: 'Task';
  descriptionMessageId?: Maybe<Scalars['String']>;
  experience: Scalars['Int'];
  factionName?: Maybe<Scalars['String']>;
  failMessageId?: Maybe<Scalars['String']>;
  finishRewards?: Maybe<TaskRewards>;
  id?: Maybe<Scalars['ID']>;
  map?: Maybe<Map>;
  minPlayerLevel?: Maybe<Scalars['Int']>;
  name: Scalars['String'];
  neededKeys?: Maybe<Array<Maybe<TaskKey>>>;
  normalizedName: Scalars['String'];
  objectives: Array<Maybe<TaskObjective>>;
  startMessageId?: Maybe<Scalars['String']>;
  startRewards?: Maybe<TaskRewards>;
  successMessageId?: Maybe<Scalars['String']>;
  tarkovDataId?: Maybe<Scalars['Int']>;
  taskRequirements: Array<Maybe<TaskStatusRequirement>>;
  trader: Trader;
  traderLevelRequirements: Array<Maybe<RequirementTrader>>;
  wikiLink?: Maybe<Scalars['String']>;
};
// ...
export type Trader = {
  __typename?: 'Trader';
  /** barters and cashOffers are only available via the traders query. */
  barters: Array<Maybe<Barter>>;
  cashOffers: Array<Maybe<TraderCashOffer>>;
  currency: Item;
  description?: Maybe<Scalars['String']>;
  discount: Scalars['Float'];
  id: Scalars['ID'];
  levels: Array<TraderLevel>;
  name: Scalars['String'];
  normalizedName: Scalars['String'];
  resetTime?: Maybe<Scalars['String']>;
  tarkovDataId?: Maybe<Scalars['Int']>;
};
// ...
export type TraderPrice = {
  __typename?: 'TraderPrice';
  /** @deprecated Use item.buyFor instead. */
  currency: Scalars['String'];
  /** @deprecated Use item.buyFor instead. */
  price: Scalars['Int'];
  /** @deprecated Use item.buyFor instead. */
  priceRUB: Scalars['Int'];
  /** @deprecated Use item.buyFor instead. */
  trader: Trader;
};
// ...
export type Vendor = {
  name: Scalars['String'];
  normalizedName: Scalars['String'];
};

export type HistoricalPricePoint = {
  __typename?: 'historicalPricePoint';
  price?: Maybe<Scalars['Int']>;
  timestamp?: Maybe<Scalars['String']>;
};